fix: Broken test due to updated parameters
